Refactor out the map classes into the lib area.
